Fix race condition.
Two functions, finishRequest and headersReceived, both delegated to async requestUpdated, but they weren't waiting on the promise to finish. This created a race condition where requests were being considered "completed" before all the headers had been copied over, and sent to the server without headers.

This change fixes those race conditions, but I'm still a bit concerned about the possibility of race conditions from the primary listeners finishing out of order. For example, is it possible that the handler for onCompleted might execute before the handler for onResponseStarted was done? I haven't been able to make that happen so far but it's something to keep in mind.