Record user actions in iframes.
Previously, we were unable to see clicks etc that happened in frames. With this change, the content script now runs in every iframe also, and we record user actions that happen there with a special pseudo-selector starting with iframe[name="name"], so that the player can then resolve which frame to run the step in.