Version Compatibility Fixes (#21)
* Check the version when injecting scripts.

This was causing a problem with recording clicks in Chrome with the v3 manifest, which I didn't notice earlier because navigations were being recorded normally.

* Return true to avoid Firefox complaints.

Returning a complicated object like the finder here made Firefox bomb with a complaint that the "result is non-structured-clonable data". Since we don't actually use the direct result of this function, I made it return true instead.

* Return true after adding listeners.

I'm not sure if this one is strictly necessary in Firefox or not, but I added it in the course of debugging, and it seems safer to explicitly return something.

* Revise the version and split manifests.

For a while, it looks like we'll be releasing v3 to Chrome and v2 to Firefox, so I included the version in both file names to make the choice explicit.