Inject content scripts when a sub-frame navigates.
A user has an app that uses a frameset, and clicking in one frame causes navigations to happen in another frame. We were only registering the content scripts in all frames when the tab navigates, so after a navigation within a sub-frame the content scripts weren't present, and recording failed. With this change we'll watch for navigations in those sub-frames and register the content scripts even if the top-level page has not changed.