V19 (#22)
* Increment the version.

* Upgrade the polyfill.

* Run the recording scripts in all frames.

A user reported that Firefox wasn't recording things that happened in frames. Chrome didn't have the problem. Running the scripts in all frames seems to do the trick.

* Include the frame selector with text selectors too.

If you clicked on something in a frame, and the text selector preference was enabled, it wasn't prepending the frame selector.

* L-232 Replace the radio button with a record indicator in popup.html.

* Inject content scripts when a sub-frame navigates.

A user has an app that uses a frameset, and clicking in one frame causes navigations to happen in another frame. We were only registering the content scripts in all frames when the tab navigates, so after a navigation within a sub-frame the content scripts weren't present, and recording failed. With this change we'll watch for navigations in those sub-frames and register the content scripts even if the top-level page has not changed.

* Fixed record indicator not showing in subframes

Co-authored-by: Konstantin Dolbilov <konfrontend@gmail.com>