Make navigate events more consistent.
We'll now automatically create a navigate event for the initially loaded URL, and afterwards only when an explicit navigation is done by the user typing a different URL into the bar. We won't create navigate events for other types of navigations since they are typically caused by the user clicking a link or button, and that's captured as a separate event anyway.