V20 (#23)
* Increment the version.

* Make content scripts able to run multiple times.

If the content script gets injected more than once, we want it to silently proceed, instead of possibly bombing or adding two of something. This is mainly so we don't have to worry about race conditions when deciding to inject them.

* Inject content scripts every time no matter what.

A customer reported that the extension wasn't able to record his site, and this led to me discovering a strange race condition where if the navigationCommitted event fired BEFORE the onUpdatedTab event, the content scripts failed to load into the parent frame. I think this only happened because of the way his site was structured with nested frames. Now, we'll just inject content scripts every time no matter what, since the content scripts themselves can safely be run again.

* Make navigate events more consistent.

We'll now automatically create a navigate event for the initially loaded URL, and afterwards only when an explicit navigation is done by the user typing a different URL into the bar. We won't create navigate events for other types of navigations since they are typically caused by the user clicking a link or button, and that's captured as a separate event anyway.